+++ curves/g_asuivre/g_as2.crv
@@ -1,0 +1,4 @@
+m=0.166341:0.000000;0.776908:0.236791;0.933464:1.000000
+r=0.000000:0.000000;0.949119:1.000000
+g=0.000000:0.000000;0.661448:0.575342;1.000000:1.000000
+b=0.056751:0.000000;1.000000:1.000000
