+++ database/ep01/ep01_k.ini
@@ -781,6 +781,14 @@ sharpen="""
 
 
 [filters_asuivre]
+id_pre_upscale=4
+pre_upscale_ffmpeg=""
+
+id_upscale=4
+upscale="""
+    scale=width_upscale:height_upscale:bicubic
+    """
+
 id_sharpen=14
 sharpen="""
     fastNlMeansDenoisingColored=5:5:7:21,
