+++ known_issues.md
@@ -1,14 +1,11 @@
 # video_editor
 - After saving a new RGB curves from current (save as), the previous is not restored and is considered as modified
 - When selecting curves on multiple shots, saving does not work (save all)
-- list modifications on the "save before closing" dialog
-- custom geometry is not verified for g_debut, g_fin (and will not work because the section shall be the src ed/ep and not the dst ed/ep)
-- geometry shall include ep no. required for g_debut
+- crop is not working for g_debut, g_fin
 
 # merge_stabilize
 - does not work
 
 # scripts
 - geometry is not working
-- does not save image after rgb (except for frame study?)
-- final image shall be saved in the dst episode and not in the src because of geometry which can be different
+- do not save image after rgb (except for frame study?)
