+++ tools/common/window_common.py
@@ -30,6 +30,8 @@ from PySide6.QtWidgets import (
 from common.sylesheet import set_widget_stylesheet
 from utils.common import FPS
 
+PAINTER_MARGIN_LEFT = 20
+PAINTER_MARGIN_TOP = 20
 
 class Window_common(QMainWindow):
 
@@ -72,7 +74,10 @@ class Window_common(QMainWindow):
         self.current_editor = ''
         self.current_widget = self.current_editor
 
-        self.show_side = 'top'
+        # This is used  when the screen height is <= 1080 to display
+        # the bottom side
+        self.display_position_y = 0
+        self.display_height = QApplication.screens()[0].size().height()
 
         self.current_frame_index = -1
         self.playing_frame_start_no = 0
@@ -86,9 +91,6 @@ class Window_common(QMainWindow):
 
         # Connect signals coming from model
         self.model.signal_shotlist_modified[dict].connect(self.event_shotlist_modified)
-
-
-
 
 
 
@@ -174,8 +176,11 @@ class Window_common(QMainWindow):
                 message_box = QMessageBox()
                 message_box.setIcon(QMessageBox.Warning)
                 message_box.setWindowTitle("Save before closing?")
-                message_box.setText("Some modifications have not been change.");
-                message_box.setInformativeText("Do you want to close before saving?");
+                text = "Some modifications have not been saved:"
+                for s in self.model.get_modified_db():
+                    text += "\n  - %s" % (s)
+                message_box.setText(text)
+                message_box.setInformativeText("Do you want to close before saving?")
                 message_box.setStandardButtons(QMessageBox.Save | QMessageBox.Discard | QMessageBox.Cancel)
                 message_box.setDefaultButton(QMessageBox.Save)
                 set_widget_stylesheet(message_box)
@@ -234,6 +239,7 @@ class Window_common(QMainWindow):
 
 
     def event_shotlist_modified(self, shotlist):
+        # episode/part has been changed
         enabled = True if len(shotlist['shots']) > 0 else False
         for w_str in self.model.get_widget_list():
             log.info("%s: set enabled: %s" % (self.widgets[w_str].objectName(), 'true' if enabled else 'false'))
@@ -247,6 +253,26 @@ class Window_common(QMainWindow):
             try: self.widgets[w_str].set_widget_enabled(enabled)
             except: pass
 
+
+    def switch_display_side(self):
+        if self.display_height > 1080:
+            return
+        self.event_screen_position_changed('switch')
+        self.repaint()
+
+
+    def event_screen_position_changed(self, side):
+        # log.info("change side to %s" % (side))
+        if side == 'switch':
+            new_side = 'bottom' if self.display_position_y == 0 else 'top'
+        else:
+            new_side = side
+
+        if new_side == 'bottom':
+            self.display_position_y = 1152 - 1080 + 2*PAINTER_MARGIN_TOP
+        else:
+            self.display_position_y = 0
+        self.repaint()
 
 
     def get_current_widget(self):
@@ -295,6 +321,10 @@ class Window_common(QMainWindow):
                 break
 
 
+    def event_selected_shots_changed(self, selection):
+        self.event_preview_options_changed('selection')
+
+
     def event_preview_options_changed(self, widget):
         # log.info("change preview: editor: %s" % (self.current_editor))
         log.info("change preview: editor: %s" % (widget))
@@ -302,8 +332,6 @@ class Window_common(QMainWindow):
         for e, w in self.widgets.items():
             preview_options.update({e: w.get_preview_options()})
         self.signal_preview_options_changed.emit(preview_options)
-
-
 
 
     def event_ready_to_play(self, playlist_properties):
@@ -372,12 +400,6 @@ class Window_common(QMainWindow):
             self.widget_controls.set_playing_frame_properties(self.current_frame_index)
             f = self.model.get_frame(self.current_frame_index + self.playing_frame_start_no)
             self.display_frame(f)
-
-
-
-    def event_upper_lower_preview_changed(self, side:str):
-        self.show_side = side
-        self.repaint()
 
 
     def event_right_click(self, qpoint):
